throw Error("Image not in queue") if img is not found

+++ picgen.js
@@ -14,6 +14,7 @@ router.get('/', function (req, res) {
     console.log(req.query);
 
     let sendFile = function (img) {
+        if(img === undefined) throw new Error("Image not in queue");
         let file = path.join(__dirname + "/" + img);
         res.sendFile(file, () => fs.unlink(file, () => console.log("send dele: " + img)));
     };
